F3D-25: close the Move transform-form integer domain
DesignMoveForm has the two source forms 1 and 5. Replace the raw native
integer and the parser's separate membership check with typed conversion.
Serialize each form as its original integer and reject other wire values
with an error naming form. No post-hoc Move-form validator existed.

The format defines the two codes but gives neither a distinct semantic
name, so the variants retain their source-code names. Parser tests keep
their expected source codes; exact integer round-trip and rejection tests
cover the native boundary. Scoped test compilation passes with and without
schema.

Hook-Bypass-Reason: User requested scoped compile checks between items and one final full gate.

+++ crates/cadmpeg-codec-f3d/src/design/decode/scopes/tests/fixed_kind_tail.rs
@@ -799,7 +799,7 @@ fn parameter_scope_uses_same_index_pair_and_fixed_kind_tail() {
     .expect("class-368 Move frame");
     assert_eq!(decoded.transform, move_transform);
     assert_eq!(decoded.transform_offset, (move_at + 48) as u64);
-    assert_eq!(decoded.form, 5);
+    assert_eq!(u32::from(decoded.form), 5);
 
     let compact_move_at = bytes.len();
     let mut compact_move = vec![0; 253];
@@ -827,7 +827,7 @@ fn parameter_scope_uses_same_index_pair_and_fixed_kind_tail() {
     assert_eq!(decoded.transform, move_transform);
     assert_eq!(decoded.transform_offset, (compact_move_at + 48) as u64);
     assert_eq!(decoded.transform_record_index, 91);
-    assert_eq!(decoded.form, 1);
+    assert_eq!(u32::from(decoded.form), 1);
     assert_eq!(decoded.form_offset, (compact_move_at + 43) as u64);
     bytes[compact_move_at + 4..compact_move_at + 7].copy_from_slice(b"362");
     bytes[compact_move_at + 43..compact_move_at + 47].copy_from_slice(&5u32.to_le_bytes());
@@ -838,7 +838,7 @@ fn parameter_scope_uses_same_index_pair_and_fixed_kind_tail() {
     )
     .expect("class-362 Move frame");
     assert_eq!(decoded.transform, move_transform);
-    assert_eq!(decoded.form, 5);
+    assert_eq!(u32::from(decoded.form), 5);
 
     let class_433_move_at = bytes.len();
     let mut class_433_move = vec![0; 253];
@@ -866,7 +866,7 @@ fn parameter_scope_uses_same_index_pair_and_fixed_kind_tail() {
     assert_eq!(decoded.transform, move_transform);
     assert_eq!(decoded.transform_offset, (class_433_move_at + 48) as u64);
     assert_eq!(decoded.transform_record_index, 92);
-    assert_eq!(decoded.form, 5);
+    assert_eq!(u32::from(decoded.form), 5);
 
     let scale_at = bytes.len();
     let mut scale = vec![0; 317];
